chore(erc4626): new cfContract protocol format
contractInfo :=
  baseContractInfo
  [protocolExtensions]

baseContractInfo :=
  name: string
  symbol: string
  standard: string
  decimals: vuint
  createdInBlock: vuint
  destructedInBlock: vuint

protocolExtensions :=
  extensionHeader: vuint
  extensionCount: vuint
  repeated extensionCount times:
    protocolId: vuint
    payloadLength: vuint
    payload: bytes
