+++ db/rocksdb_ethereumtype.go
@@ -9,7 +9,6 @@ import (
 	"sync"
 	"time"
 
-	vlq "github.com/bsm/go-vlq"
 	"github.com/golang/glog"
 	"github.com/juju/errors"
 	"github.com/linxGnu/grocksdb"
@@ -958,180 +957,6 @@ func (d *RocksDB) storeInternalDataEthereumType(wb *grocksdb.WriteBatch, blockTx
 		if blockTx.internalData != nil {
 			wb.PutCF(d.cfh[cfInternalData], blockTx.btxID, packEthInternalData(blockTx.internalData))
 		}
-	}
-	return nil
-}
-
-var cachedContracts = newContractInfoLRU(cachedContractsLRUMaxSize)
-
-func packContractInfo(contractInfo *bchain.ContractInfo) []byte {
-	buf := packString(contractInfo.Name)
-	buf = append(buf, packString(contractInfo.Symbol)...)
-	buf = append(buf, packString(string(contractInfo.Standard))...)
-	varBuf := make([]byte, vlq.MaxLen64)
-	l := packVaruint(uint(contractInfo.Decimals), varBuf)
-	buf = append(buf, varBuf[:l]...)
-	l = packVaruint(uint(contractInfo.CreatedInBlock), varBuf)
-	buf = append(buf, varBuf[:l]...)
-	l = packVaruint(uint(contractInfo.DestructedInBlock), varBuf)
-	buf = append(buf, varBuf[:l]...)
-	// Trailing fields below were appended after the original layout. Records written
-	// by older versions stop earlier, and unpack reads back zero values for them.
-	// The flags varuint packs multiple bits to keep the on-disk shape stable as we
-	// add more boolean signals; bit 0 is IsErc4626. Bit 1 was used by an earlier
-	// persistent negative-probe scheme that has since been replaced by an in-memory
-	// cache; it is no longer written and is silently ignored when read.
-	var flags uint
-	if contractInfo.IsErc4626 {
-		flags |= 1
-	}
-	l = packVaruint(flags, varBuf)
-	buf = append(buf, varBuf[:l]...)
-	buf = append(buf, packString(contractInfo.Erc4626AssetContract)...)
-	return buf
-}
-
-func unpackContractInfo(buf []byte) (*bchain.ContractInfo, error) {
-	var contractInfo bchain.ContractInfo
-	var s string
-	var l int
-	var ui uint
-	contractInfo.Name, l = unpackString(buf)
-	buf = buf[l:]
-	contractInfo.Symbol, l = unpackString(buf)
-	buf = buf[l:]
-	s, l = unpackString(buf)
-	contractInfo.Standard = bchain.TokenStandardName(s)
-	contractInfo.Type = bchain.TokenStandardName(s)
-	buf = buf[l:]
-	ui, l = unpackVaruint(buf)
-	contractInfo.Decimals = int(ui)
-	buf = buf[l:]
-	ui, l = unpackVaruint(buf)
-	contractInfo.CreatedInBlock = uint32(ui)
-	buf = buf[l:]
-	ui, l = unpackVaruint(buf)
-	contractInfo.DestructedInBlock = uint32(ui)
-	buf = buf[l:]
-	// Older records may not have these trailing fields; vlq decode of an empty buffer
-	// returns 0, and unpackString reads "" when its length-prefix is 0.
-	ui, l = unpackVaruint(buf)
-	contractInfo.IsErc4626 = ui&1 != 0
-	buf = buf[l:]
-	contractInfo.Erc4626AssetContract, _ = unpackString(buf)
-	return &contractInfo, nil
-}
-
-func (d *RocksDB) GetContractInfoForAddress(address string) (*bchain.ContractInfo, error) {
-	contract, err := d.chainParser.GetAddrDescFromAddress(address)
-	if err != nil || contract == nil {
-		return nil, err
-	}
-	return d.GetContractInfo(contract, "")
-}
-
-// GetContractInfo gets contract from cache or DB and possibly updates the standard from standardFromContext
-// it is hard to guess the standard of the contract using API, it is easier to set it the first time the contract is processed in a tx
-func (d *RocksDB) GetContractInfo(contract bchain.AddressDescriptor, standardFromContext bchain.TokenStandardName) (*bchain.ContractInfo, error) {
-	cacheKey := string(contract)
-	contractInfo, found := cachedContracts.get(cacheKey)
-	if !found {
-		val, err := d.db.GetCF(d.ro, d.cfh[cfContracts], contract)
-		if err != nil {
-			return nil, err
-		}
-		defer val.Free()
-		buf := val.Data()
-		if len(buf) == 0 {
-			return nil, nil
-		}
-		contractInfo, _ = unpackContractInfo(buf)
-		addresses, _, _ := d.chainParser.GetAddressesFromAddrDesc(contract)
-		if len(addresses) > 0 {
-			contractInfo.Contract = addresses[0]
-		}
-		// if the standard is specified and stored contractInfo has unknown standard, set and store it
-		if standardFromContext != bchain.UnknownTokenStandard && contractInfo.Standard == bchain.UnknownTokenStandard {
-			contractInfo.Standard = standardFromContext
-			contractInfo.Type = standardFromContext
-			err = d.db.PutCF(d.wo, d.cfh[cfContracts], contract, packContractInfo(contractInfo))
-			if err != nil {
-				return nil, err
-			}
-		}
-		cachedContracts.add(cacheKey, contractInfo)
-	}
-	return contractInfo, nil
-}
-
-// SetContractInfoErc4626Vault persists the cached ERC4626 invariants for a
-// detected vault: marks IsErc4626=true and stores the underlying asset address.
-// If the row does not yet exist, this is a no-op - the contractInfo path will
-// have triggered a lazy ContractInfo fetch separately, so the row will be
-// present by the time we reach here in normal flow. Idempotent.
-func (d *RocksDB) SetContractInfoErc4626Vault(address string, assetContract string) error {
-	contract, err := d.chainParser.GetAddrDescFromAddress(address)
-	if err != nil || contract == nil {
-		return err
-	}
-	contractInfo, err := d.GetContractInfo(contract, "")
-	if err != nil {
-		return err
-	}
-	if contractInfo == nil {
-		return nil
-	}
-	changed := false
-	if !contractInfo.IsErc4626 {
-		contractInfo.IsErc4626 = true
-		changed = true
-	}
-	if contractInfo.Erc4626AssetContract != assetContract {
-		contractInfo.Erc4626AssetContract = assetContract
-		changed = true
-	}
-	if !changed {
-		return nil
-	}
-	if err := d.db.PutCF(d.wo, d.cfh[cfContracts], contract, packContractInfo(contractInfo)); err != nil {
-		return err
-	}
-	cachedContracts.add(string(contract), contractInfo)
-	return nil
-}
-
-// StoreContractInfo stores contractInfo in DB
-// if CreatedInBlock==0 and DestructedInBlock!=0, it is evaluated as a destruction of a contract, the contract info is updated
-// in all other cases the contractInfo overwrites previously stored data in DB (however it should not really happen as contract is created only once)
-func (d *RocksDB) StoreContractInfo(contractInfo *bchain.ContractInfo) error {
-	wb := grocksdb.NewWriteBatch()
-	defer wb.Destroy()
-	if err := d.storeContractInfo(wb, contractInfo); err != nil {
-		return err
-	}
-	return d.WriteBatch(wb)
-}
-
-func (d *RocksDB) storeContractInfo(wb *grocksdb.WriteBatch, contractInfo *bchain.ContractInfo) error {
-	if contractInfo.Contract != "" {
-		key, err := d.chainParser.GetAddrDescFromAddress(contractInfo.Contract)
-		if err != nil {
-			return err
-		}
-		if contractInfo.CreatedInBlock == 0 && contractInfo.DestructedInBlock != 0 {
-			storedCI, err := d.GetContractInfo(key, "")
-			if err != nil {
-				return err
-			}
-			if storedCI == nil {
-				return nil
-			}
-			storedCI.DestructedInBlock = contractInfo.DestructedInBlock
-			contractInfo = storedCI
-		}
-		wb.PutCF(d.cfh[cfContracts], key, packContractInfo(contractInfo))
-		cacheKey := string(key)
-		cachedContracts.delete(cacheKey)
 	}
 	return nil
 }
